refactor(oohelperd): better distnguish different helpers
Part of https://github.com/ooni/probe/issues/1733
