+++ internal/cmd/oohelperd/internal/webconnectivity/fake_test.go
@@ -1,4 +1,4 @@
-package internal
+package webconnectivity
 
 import (
 	"context"
