+++ internal/cmd/oohelperd/internal/webconnectivity/webconnectivity.go
@@ -1,4 +1,4 @@
-package internal
+package webconnectivity
 
 import (
 	"encoding/json"
