+++ internal/cmd/oohelperd/internal/websteps/explore_test.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"net/http"
@@ -56,6 +56,7 @@ func TestExploreSuccessWithH3(t *testing.T) {
 
 func TestGetSuccess(t *testing.T) {
 	u, err := url.Parse("https://example.com")
+	runtimex.PanicOnError(err, "url.Parse failed for clearly good URL")
 	resp, err := explorer.get(u, nil)
 	if err != nil {
 		t.Fatal("unexpected error")
@@ -72,6 +73,7 @@ func TestGetSuccess(t *testing.T) {
 
 func TestGetFailure(t *testing.T) {
 	u, err := url.Parse("https://example.example")
+	runtimex.PanicOnError(err, "url.Parse failed for clearly good URL")
 	resp, err := explorer.get(u, nil)
 	if err == nil {
 		t.Fatal("expected an error here")
@@ -83,6 +85,7 @@ func TestGetFailure(t *testing.T) {
 
 func TestGetH3Success(t *testing.T) {
 	u, err := url.Parse("https://www.google.com")
+	runtimex.PanicOnError(err, "url.Parse failed for clearly good URL")
 	h3u := &h3URL{URL: u, proto: "h3"}
 	resp, err := explorer.getH3(h3u, nil)
 	if err != nil {
@@ -100,6 +103,7 @@ func TestGetH3Success(t *testing.T) {
 
 func TestGetH3Failure(t *testing.T) {
 	u, err := url.Parse("https://www.google.google")
+	runtimex.PanicOnError(err, "url.Parse failed for clearly good URL")
 	h3u := &h3URL{URL: u, proto: "h3"}
 	resp, err := explorer.getH3(h3u, nil)
 	if err == nil {
