+++ internal/cmd/oohelperd/internal/websteps/factory.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"context"
