+++ internal/cmd/oohelperd/internal/websteps/h3.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"errors"
