+++ internal/cmd/oohelperd/internal/websteps/h3_test.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"net/http"
