+++ internal/cmd/oohelperd/internal/websteps/initialchecks.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"context"
