+++ internal/cmd/oohelperd/internal/websteps/initialchecks_test.go
@@ -1,4 +1,4 @@
-package nwcth
+package websteps
 
 import (
 	"testing"
